Implements responding to ApiVersions requests
Additionally this introduces some major refactoring for better
separation of concerns and extensibility. In particular:

- low level tcp server interactions are extracted into a Kundera
class in a broker module. this module exposes a createBroker function
which is called in the entrypoint

- to avoid repeating interfaces and to make dependency explicit, the
net.Server instance is still returned by createBroker, however it does
include a monkeypatched method `handle` accepting an ApiKey and a handler
callback which will be invoked with a request and response parameter. this
is used to register handler functions responsible for providing responses
to the requests of type ApiKey.

- serializer entities encapsulate the logic of writing primitives to binary
buffers and message_body entities encapsulate the detailed structure of
api responses and requests. this allows for an expressive definition of the
fields contained in the requests/responses following the Kafka Wire Protocol
without being littered with the actual transformations into binary suitable
to put on the wire. userland code within the broker should be able to work
entirely with message_body entities and the message_bodies should be the
only place where detailed knowledge of the wire protocol is required. That is,
userland code should largely be able to work with javascript data structures
that logically map to the wire protocol primitives.